+++ src/main/java/net/snowflake/client/jdbc/RestRequest.java
@@ -17,8 +17,6 @@ import net.snowflake.common.core.SqlState;
 
 import java.net.URI;
 
-import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpRequestBase;
 import org.apache.http.client.utils.URIBuilder;